# tutorflow-server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_tutorflow_tutorflow-server_internal_pkg_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.ErrorInfo'
      message:
        type: string
      meta:
        $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Meta'
      success:
        type: boolean
    type: object
...
  github_com_tutorflow_tutorflow-server_internal_usecase_course.CreateInput:
    properties:
      category_ids:
        items:
          type: string
        type: array
      description:
        type: string
      language:
        type: string
      level:
        enum:
        - beginner
        - intermediate
        - advanced
        type: string
      price:
        minimum: 0
        type: number
      requirements:
        items:
          type: string
        type: array
      short_description:
        maxLength: 500
        type: string
      title:
        maxLength: 255
        minLength: 5
        type: string
      what_you_learn:
        items:
          type: string
        type: array
    required:
    - level
    - title
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@tutorflow.com
    name: API Support
    url: http://www.tutorflow.com/support
  description: This is the backend API for TutorFlow Learning Management System.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: TutorFlow LMS API
  version: "1.0"
paths:
...
  /announcements/course/{courseId}:
    get:
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get course announcements
      tags:
      - Announcements
...
  /courses:
    get:
      parameters:
      - description: Search term
        in: query
        name: search
        type: string
      - description: Level filter
        in: query
        name: level
        type: string
      - description: Category ID
        in: query
        name: category_id
        type: string
      - description: Minimum price
        in: query
        name: min_price
        type: number
      - description: Maximum price
        in: query
        name: max_price
        type: number
      - description: 'Sort by: created_at, price, rating, students'
        in: query
        name: sort_by
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
      summary: List courses
      tags:
      - Courses
    post:
      consumes:
      - application/json
      parameters:
      - description: Course data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_usecase_course.CreateInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Course'
              type: object
      security:
      - BearerAuth: []
      summary: Create course
      tags:
      - Courses
  /courses/{courseId}/modules:
    get:
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
      summary: List modules for a course
      tags:
      - Modules
    post:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      - description: Module data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_usecase_course.CreateModuleInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Module'
              type: object
      security:
      - BearerAuth: []
      summary: Create a module
      tags:
      - Modules
  /courses/{courseId}/modules/{moduleId}:
    delete:
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      - description: Module ID
        in: path
        name: moduleId
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a module
      tags:
      - Modules
    put:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      - description: Module ID
        in: path
        name: moduleId
        required: true
        type: string
      - description: Module data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_usecase_course.UpdateModuleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Module'
              type: object
      security:
      - BearerAuth: []
      summary: Update a module
      tags:
      - Modules
  /courses/{courseId}/modules/reorder:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      - description: Module IDs in order
        in: body
        name: request
        required: true
        schema:
          properties:
            module_ids:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Reorder modules
      tags:
      - Modules
  /courses/{id}:
    delete:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete course
      tags:
      - Courses
    put:
      consumes:
      - application/json
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      - description: Course data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_usecase_course.UpdateInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Course'
              type: object
      security:
      - BearerAuth: []
      summary: Update course
      tags:
      - Courses
  /courses/{id}/archive:
    patch:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Archive course
      tags:
      - Courses
  /courses/{id}/curriculum:
    get:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
      summary: Get course curriculum
      tags:
      - Courses
  /courses/{id}/publish:
    patch:
      parameters:
      - description: Course ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Publish course
      tags:
      - Courses
  /courses/{idOrSlug}:
    get:
      parameters:
      - description: Course ID or slug
        in: path
        name: idOrSlug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Course'
              type: object
      summary: Get course
      tags:
      - Courses
  /courses/categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Category'
                  type: array
              type: object
      summary: List all categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      parameters:
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_usecase_course.CreateCategoryInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Category'
              type: object
      security:
      - BearerAuth: []
      summary: Create a category (admin)
      tags:
      - Categories
  /courses/categories/{id}:
    delete:
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete a category (admin)
      tags:
      - Categories
    put:
      consumes:
      - application/json
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      - description: Category data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_usecase_course.CreateCategoryInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_domain.Category'
              type: object
      security:
      - BearerAuth: []
      summary: Update a category (admin)
      tags:
      - Categories
...
  /courses/my:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get my courses (as instructor)
      tags:
      - Courses
...
  /discussions/course/{courseId}:
    get:
      parameters:
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_tutorflow_tutorflow-server_internal_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: Get discussions for a course
      tags:
      - Discussions
...
  /learning-paths/{id}/courses/{courseId}:
    delete:
      parameters:
      - description: Path ID
        in: path
        name: id
        required: true
        type: string
      - description: Course ID
        in: path
        name: courseId
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Remove a course from a learning path
      tags:
      - Learning Paths
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and then your API key.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"